Can install Blender multiple times without error

+++ Editor/Scripts/DCCIntegration/BlenderIntegrator.cs
@@ -39,22 +39,27 @@ internal class BlenderIntegrator : BaseDCCIntegrator {
             return false;
         }
         
-        //script
+        //Prepare install script
         string installScriptFileName = $"InstallBlenderPlugin_{ver}.py";
         string templatePath = Path.Combine(MeshSyncEditorConstants.DCC_INSTALL_SCRIPTS_PATH,installScriptFileName );
         if (!File.Exists(templatePath)) {
             return false;
         }
-
-       
+        
         //Replace the path in the template with actual path.
         string installScriptFormat = File.ReadAllText(templatePath);
         string installScript = String.Format(installScriptFormat,pluginFile);
         string installScriptPath = Path.Combine(tempPath, installScriptFileName);
         File.WriteAllText(installScriptPath, installScript);
         
+        //Prepare remove script to remove old plugin
+        string uninstallScriptFilename = $"UninstallBlenderPlugin_{ver}.py";
+        string uninstallScript = Path.Combine(MeshSyncEditorConstants.DCC_INSTALL_SCRIPTS_PATH,uninstallScriptFilename );
+        if (!File.Exists(uninstallScript)) {
+            return false;
+        }
       
-        bool setupSuccessful = SetupAutoLoadPlugin(dccToolInfo.AppPath, installScriptPath);
+        bool setupSuccessful = SetupAutoLoadPlugin(dccToolInfo.AppPath, uninstallScript, installScriptPath);
 
         //Cleanup
         File.Delete(installScriptPath);
@@ -76,7 +81,7 @@ internal class BlenderIntegrator : BaseDCCIntegrator {
 //----------------------------------------------------------------------------------------------------------------------
 
     
-    bool SetupAutoLoadPlugin(string appPath, string installScriptPath) {
+    bool SetupAutoLoadPlugin(string appPath, string uninstallScriptPath, string installScriptPath) {
 
         try {
             if (!System.IO.File.Exists(appPath)) {
@@ -84,6 +89,7 @@ internal class BlenderIntegrator : BaseDCCIntegrator {
                 return false;
             }
 
+            //Try to uninstall first. The uninstallation may have error messages, but they can be ignored
             System.Diagnostics.Process process = new System.Diagnostics.Process {
                 StartInfo = {
                     FileName = appPath,
@@ -91,18 +97,22 @@ internal class BlenderIntegrator : BaseDCCIntegrator {
                     // CreateNoWindow = true,
                     UseShellExecute = false,
                     RedirectStandardError = true,
-                    Arguments = $"-b -P {installScriptPath}"         //Execute batch mode
+                    Arguments = $"-b -P {uninstallScriptPath}"         //Execute batch mode
                 },
                 EnableRaisingEvents = true
-            };
+            };            
             process.Start();
-            
-
-            string stderr = process.StandardError.ReadToEnd();
             process.WaitForExit();
 
-            if (stderr.Contains("Error")) {
-                Debug.LogError($"Installation error. {stderr}");
+
+            //Install
+            process.StartInfo.Arguments = $"-b -P {installScriptPath}";
+            process.Start();
+            process.WaitForExit();
+            string stderr = process.StandardError.ReadToEnd();
+            
+            if (stderr.Contains("PermissionError")) {
+                Debug.LogError($"[MeshSync] Installation error. {stderr}");
                 return false;
             }
